kustomize-playground/pyproject.toml
[project]
name = "kustomize-playground"
version = "0.1.0"
description = "A web-based Kustomize playground for testing patches"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "jinja2==3.1.3",
    "pyyaml==6.0.1",
    "python-multipart==0.0.9"
]

[dependency-groups]
dev = [
    "truefoundry==0.6.1",
    "pre-commit==3.6.2",
    "ruff==0.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
package = false

[tool.ruff]
target-version = "py311"
exclude = [
  ".venv",
  "venv",
  "repos",
]

[tool.ruff.lint]
select = [
  "E",     # pycodestyle errors
  "W",     # pycodestyle warnings
  "F",     # pyflakes
  "C",     # flake8-comprehensions
  "B",     # flake8-bugbear
  "Q",     # flake8-quotes
  "I",
  "T10",
  "ASYNC",
  "NPY201",
]
ignore = ["E501"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
